exkml v0.2.5 Exkml

Documentation for Exkml.

Functions

ack(from, ref)

emit(state)

events!(binstream, chunk_size \\ 2048)

Get events sent directly to the calling process

Returns a unique reference you can use to match on, ref = events!(kml_byte_stream)

Event will be one of {:placemarks, ^ref, from, placemarks} Where placemarks are a list of marks

When you have done what you need to do, send an acknowledgement. The ack contains the ref ref, and is a message back to the from process to make it start parsing the next batch

Like this

  Exkml.ack(from, ref)

{:done, ^ref} When parsing is done {:error, ^ref, ^pid, event} An error with event being the last SAX event

on_event(_, event, state)

pluck_attribute(attributes, name, value)

pop_event(state)

pop_geom(state)

pop_geom(state, kind)

push_event(state, event)

push_geom(state, geom)

put_attribute(s, name, value)

put_error(state, reason)

put_in_placemark(pm, geom)

put_line(state, text)

put_point(state, text)

setup(binstream, chunk_size, ref)

stage(binstream, chunk_size \\ 2048)

str_to_point(point_str)

stream!(binstream, chunk_size \\ 2048)

Get a stream of placemarks